Django Yeni Sitemiz/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-ckeditor==6.7.2
django-js-asset==3.1.2
Pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1